% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{render_report_reprogram_seq}
\alias{render_report_reprogram_seq}
\title{Generate the report with the use of reprogram-seq data}
\usage{
render_report_reprogram_seq(
  include_other_packages = FALSE,
  regulon = NULL,
  treatments = list(GATA6 = NULL, NKX2.1 = NULL),
  output_file_sumary = NULL,
  cluster_column = "Clusters",
  regulon_script = "",
  data_file_paths = "",
  temp_folder = "",
  conda_exe = "",
  virtual_env = "",
  work_dir = "",
  output_file = "Reprogram-seq.html"
)
}
\arguments{
\item{include_other_packages}{A logical indicating whether the raport should present
benchmarking the results using other tools i.e. FigR, Pando, cellOracle and scenic+.}

\item{regulon}{A data frame being the output from epiregulon workflow. The weight
column, if present, will be disregarded.}

\item{treatments}{A named list of character vectors. Each name corresponds to
one treatment and each vector to respective control groups. Treatments and control groups
are matched to the values in the \code{hash_assignment} column from the \code{GeneExpressionMatrix}.
If the name points to the \code{NULL} object then control set of cell is created from all the groups
apart from that indicated by the element name. List names should contain transcription factors' names.
The default is set to \code{list(GATA6 = NULL, NKX2.1 = NULL)}.}

\item{cluster_column}{A character specifying the column in the \code{GeneExpressionMatrix} in which cluster information
is stored. By default is is \code{Clusters}.}

\item{regulon_script}{A full path to the R script file with the code generating the regulon. It
will not be evaluated.}

\item{output_file_summary}{A path to the .csv file to which data frame with the summary of the analysis output will be saved.}
}
\description{
Generate the report with the use of reprogram-seq data
}
