% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{render_report_reprogram_seq}
\alias{render_report_reprogram_seq}
\alias{render_report_simulation}
\title{Generate the report with the use of reprogram-seq data}
\usage{
render_report_reprogram_seq(
  output_file = "Reprogram-seq_benchmark.html",
  data_file_paths =
    "/gne/data/lab-shares/xie-lab/Sequencing_Data/2022/mapping/20220124_ReprogramSeq_Multiome/JT65_67/outs/",
  motif_db_dir = find.package("epiregulon.benchmark"),
  save_results = TRUE,
  save_path = NULL,
  ...
)

render_report_simulation(output_file = "simulation.html", ...)
}
\arguments{
\item{motif_db_dir}{Path to the direcotry with motif databases used by \code{pycistarget}. Can be downloaded from
https://resources.aertslab.org/cistarget/databases/homo_sapiens/hg38/screen/mc_v10_clust/region_based/.}

\item{treatments}{A named list of character vectors. Each name corresponds to
one treatment and each vector to respective control groups. Treatments and control groups
are matched to the values in the \code{hash_assignment} column from the \code{GeneExpressionMatrix}.
If the name points to the \code{NULL} object then control set of cell is created from all the groups
apart from that indicated by the element name. List names should contain transcription factors' names.
The default is set to \code{list(GATA6 = NULL, NKX2.1 = NULL)}.}

\item{motif_score}{A logical indicating whether motif scores should be added. If so, regulon rows with zero motif score
will have also 0 value for the weight.}
}
\description{
Generate the report with the use of reprogram-seq data
}
