% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reports.R
\name{render_report_reprogram_seq}
\alias{render_report_reprogram_seq}
\title{Generate the report with the use of reprogram-seq data}
\usage{
render_report_reprogram_seq(
  output_file = file.path(getwd(), "Reprogram-seq_benchmark.html"),
  ...
)
}
\arguments{
\item{regulon}{A data frame being the output from epiregulon workflow. The weight
column, if present, will be disregarded.}

\item{regulon_script}{A full path to the R script file with the code generating the regulon. It
will not be evaluated.}

\item{treatments}{A named list of character vectors. Each name corresponds to
one treatment and each vector to respective control groups. Treatments and control groups
are matched to the values in the \code{hash_assignment} column from the \code{GeneExpressionMatrix}.
If the name points to the \code{NULL} object then control set of cell is created from all the groups
apart from that indicated by the element name. List names should contain transcription factors' names.
The default is set to \code{list(GATA6 = NULL, NKX2.1 = NULL)}.}

\item{motif_score}{A logical indicating whether motif scores should be added. If so, regulon rows with zero motif score
will have also 0 value for the weight.}
}
\description{
Generate the report with the use of reprogram-seq data
}
